suggested changes applied
