+++ .ncurc.json
@@ -2,6 +2,8 @@
     "upgrade": true,
     "reject": [
         "react",
-        "react-dom"
+        "react-dom",
+        "@types/react",
+        "@types/react-dom"
     ]
 }
