+++ README.md
@@ -21,7 +21,7 @@ The UnoReactDraftjs component uses the state of it's father to store the markdow
 
 UnoReactDraftjs manages two states. The markdown string passed by it's father and an EditorState that the component uses to refresh the UI. The communication between these two states is unidirectional most of the time. This means that when the EditoState from UnoReactDraftjs changes it's value the markdown string updates as well, but the same doesn´t happen the other way around.
 
-If you need to update the UI state from the UnoReactDraftj's father you must implement a **RichTextEditorRef** and call it's **updateValue** method.
+If you need to update the UI state from the UnoReactDraftj's father you must implement a **RichTextEditorRef** and call it's **setText** method.
 
 ```tsx
   const draftRef = useRef<RichTextEditorRef>();
@@ -31,7 +31,7 @@ If you need to update the UI state from the UnoReactDraftj's father you must imp
       <UnoReactDraftjs markdown={markdownString} setMark={setMarkdownSting} placeholder={''} ref={draftRef} />
       <button
         onClick={() => {
-            draftRef.current.updateUI(newMarkdownString);
+            draftRef.current.setText(newMarkdownString);
         }}
       >
     </>
