+++ package.json
@@ -65,28 +65,18 @@
     "webpack": "^5.10.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^3.11.0",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-select": "3.1.0"
+    "react": "16.8.5",
+    "react-dom": "16.8.5"
   },
   "peerDependencies": {
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-select": "3.1.0"
+    "react": "-16.8.5",
+    "react-dom": "16.8.5"
   },
   "dependencies": {
     "@fluentui/react": "^7.153.4",
-    "@microsoft/applicationinsights-react-js": "^3.0.5",
-    "@microsoft/applicationinsights-web": "^2.5.10",
-    "@progress/kendo-drawing": "^1.9.3",
     "@types/draft-js": "^0.10.44",
     "draft-js": "^0.11.7",
-    "history": "^5.0.0",
-    "libphonenumber-js": "^1.9.6",
     "markdown-draft-js": "^2.2.1",
-    "tubular-common": "^5.0.1",
-    "tubular-fabric": "^3.0.1",
-    "tubular-react-common": "^3.0.1",
     "uno-react": "^1.0.16"
   },
   "jest": {
